rename namespace.py to review_app.py
