+++ hokusai/commands/__init__.py
@@ -15,7 +15,7 @@ from hokusai.commands.pull import pull
 from hokusai.commands.run import run
 from hokusai.commands.setup import setup
 from hokusai.commands.kubernetes import k8s_create, k8s_update, k8s_delete, k8s_status, k8s_copy_config
-from hokusai.commands.namespace import create_new_app_yaml, list_namespaces
+from hokusai.commands.review_app import create_new_app_yaml, list_namespaces
 from hokusai.commands.retag import retag
 from hokusai.commands.test import test
 from hokusai.commands.version import version
